refactor: types for posting
