+++ src/features/map-posting/types/geoman-minimal.d.ts
